+++ test/syntax-stringify.js
@@ -2,7 +2,7 @@ const assert = require('assert');
 const { syntax: { parse, stringify } } = require('./helpers/lib');
 const allSyntax = require('./helpers/all-syntax');
 
-describe('stringify', () => {
+describe('syntax/stringify', () => {
     it('basic test', () => {
         const actual = stringify(parse(allSyntax).ast);